CSI-923 getUserEndlessRecursion (#32)
StitchCore
NXMStitchCore - removed the specific getter name for user and token. the user/token properties are readonly in the h file and readwrite in the m file

StitchClient
calling the .user property as is exposed in the core header file.

StitchCoreApp
calling the .user property as is exposed in the core header file.

StitchClientApp
calling the .user property as is exposed in the client header file.